Unexport utils from index and lib/pmcrypto, rewrite in TS
Force importing lib/utils to also ensure that no openpgpjs code is included
