+++ lib/index.ts
@@ -1,3 +1,3 @@
 export { WorkerPool as CryptoWorkerPool } from './worker/workerPool';
 export { Api as CryptoWorkerApi } from './worker/api';
-export { VERIFICATION_STATUS, SIGNATURE_TYPES, utf8ArrayToString } from './pmcrypto'; // TODO export utility functions here, outside of worker
+export { VERIFICATION_STATUS, SIGNATURE_TYPES } from './constants';
