+++ lib/message/utils.js
@@ -11,7 +11,8 @@ import {
     CleartextMessage,
     readMessage,
     Signature,
-    readSignature
+    readSignature,
+    unarmor
 } from '../openpgp';
 import { VERIFICATION_STATUS } from '../constants';
 import { serverTime } from '../serverTime';
@@ -311,3 +312,14 @@ export async function armorBytes(value) {
     const bodyMessage = await getMessage(value);
     return readToEnd(bodyMessage.armor());
 }
+
+/**
+ * Dearmor a pgp encoded message.
+ * @param {String} input
+ * @return {Promise<Uint8Array>}
+ */
+export const stripArmor = async (input) => {
+    const { data } = await unarmor(input);
+    const bytes = await readToEnd(data);
+    return bytes;
+};
