+++ lib/pmcrypto.js
@@ -1,5 +1,4 @@
 import { setConfig } from './openpgp';
-import { encodeUtf8, decodeUtf8, decodeUtf8Base64, encodeUtf8Base64, encodeBase64, decodeBase64 } from './utils';
 
 export function init() {
     if (arguments.length) {
@@ -11,22 +10,6 @@ export function init() {
 export { updateServerTime, serverTime } from './serverTime';
 
 export { SHA256, SHA512, unsafeMD5, unsafeSHA1 } from './crypto/hash';
-
-export {
-    decodeUtf8Base64,
-    encodeUtf8Base64,
-    encodeUtf8,
-    decodeUtf8,
-    encodeBase64,
-    decodeBase64,
-    concatArrays,
-    binaryStringToArray,
-    arrayToBinaryString,
-    arrayToHexString,
-    stripArmor,
-    utf8ArrayToString,
-    stringToUtf8Array
-} from './utils';
 
 export { checkMailboxPassword } from './mail';
 
@@ -57,7 +40,8 @@ export {
     verifyMessage,
     verifyCleartextMessage,
     getCleartextMessage,
-    armorBytes
+    armorBytes,
+    stripArmor
 } from './message/utils';
 
 export { parseMail, default as processMIME } from './message/processMIME';
@@ -67,11 +51,3 @@ export { default as keyInfo, getSHA256Fingerprints } from './key/info';
 export { checkKeyStrength, keyCheck } from './key/check';
 
 export * from './constants';
-
-/* eslint-disable camelcase */
-export const encode_utf8 = encodeUtf8;
-export const decode_utf8 = decodeUtf8;
-export const encode_base64 = encodeBase64;
-export const decode_base64 = decodeBase64;
-export const encode_utf8_base64 = encodeUtf8Base64;
-export const decode_utf8_base64 = decodeUtf8Base64;
