+++ test/message/encryptMessage.spec.ts
@@ -1,13 +1,12 @@
 import { expect } from 'chai';
 // @ts-ignore missing web-stream-tools types
-import { readToEnd, ReadableStream, WritableStream, toStream } from '@openpgp/web-stream-tools';
+import { readToEnd, ReadableStream, WritableStream, toStream, WebStream } from '@openpgp/web-stream-tools';
 import { config, readMessage, CompressedDataPacket, enums } from '../../lib/openpgp';
 
-import { decryptPrivateKey, getMessage, verifyMessage, encryptMessage, decryptMessage, getSignature, stringToUtf8Array  } from '../../lib/pmcrypto';
-import type { WebStream } from '../../lib/pmcrypto';
+import { decryptPrivateKey, getMessage, verifyMessage, encryptMessage, decryptMessage, getSignature  } from '../../lib/pmcrypto';
+import { hexToUint8Array, arrayToBinaryString, stringToUtf8Array } from '../../lib/utils';
 import { testPrivateKeyLegacy } from './decryptMessageLegacy.data';
 import { VERIFICATION_STATUS } from '../../lib/constants';
-import { hexToUint8Array, arrayToBinaryString } from '../../lib/utils';
 
 const generateStreamOfData = (): { stream: WebStream<string>, data: string } => ({
     stream: new ReadableStream({ pull: (controller: WritableStream) => { for (let i = 0; i < 10000; i++ ) { controller.enqueue('string'); } controller.close() } }),
